Split out Container and SpaceView specialization from api.py
